Prepare memory map for passing between loader and kernel.

+++ shared/src/memory.rs
@@ -12,6 +12,7 @@ pub const PAGE_SIZE: usize = 1 << (PAGE_ORDER);
  * any sane configuration.
  */
 
+#[repr(u8)]
 #[derive(Clone, Copy, Debug, Eq, PartialEq)]
 pub enum MemoryStatus {
     Unknown = 0, // Memory is unvailable because it is of unknown
@@ -22,6 +23,7 @@ pub enum MemoryStatus {
     Reserved,    // Memory is unavailable because it is reserved.
 }
 
+#[repr(C, packed)]
 #[derive(Clone, Copy)]
 pub struct MemoryMapEntry {
     pub base: u64,
@@ -41,9 +43,10 @@ impl Default for MemoryMapEntry {
 
 pub const MEMORY_MAP_MAX_ENTRIES: usize = 256;
 
+#[repr(C, packed)]
 pub struct MemoryMap {
     pub entries: [MemoryMapEntry; MEMORY_MAP_MAX_ENTRIES],
-    pub num_entries: usize,
+    pub num_entries: u32,
 }
 
 impl MemoryMap {
@@ -59,7 +62,7 @@ impl MemoryMap {
             };
             i += 1;
         }
-        memory_map.num_entries = i;
+        memory_map.num_entries = i as u32;
 
         memory_map
     }
@@ -71,7 +74,7 @@ impl MemoryMap {
         let mut new_entries = [Default::default(); MEMORY_MAP_MAX_ENTRIES];
 
         for i in 0..self.num_entries {
-            let e = &self.entries[i];
+            let e = &self.entries[i as usize];
 
             if e.base >= base && e.base + e.length <= base + length {
                 // When the input range fully contains entry, delete entry.
@@ -111,7 +114,7 @@ impl MemoryMap {
         for i in 0..o {
             self.entries[i] = new_entries[i];
         }
-        self.num_entries = o;
+        self.num_entries = o as u32;
     }
 }
 
@@ -129,7 +132,7 @@ impl MemoryMap {
 pub struct FrameAllocator<'a> {
     pub cur_addr: usize,
     pub mem_map: &'a MemoryMap,
-    pub cur_map_entry: usize,
+    pub cur_map_entry: u32,
 }
 
 fn align_address(address: u64, order: u32) -> u64 {
@@ -140,13 +143,13 @@ fn align_address(address: u64, order: u32) -> u64 {
 }
 
 
-fn next_entry(mem_map: &MemoryMap, ndx: usize) -> usize {
+fn next_entry(mem_map: &MemoryMap, ndx: usize) -> u32 {
     let mut i = ndx;
     loop {
         let e = mem_map.entries[i];
         if e.status == MemoryStatus::Available
             && align_address(e.base, PAGE_ORDER) < e.base + e.length {
-            return i;
+            return i as u32;
         }
         i = i + 1;
     }
@@ -156,7 +159,7 @@ impl<'a> FrameAllocator<'a> {
     pub fn new(mem_map: &'a MemoryMap) -> FrameAllocator<'a> {
         assert!(mem_map.num_entries > 0);
         let first = next_entry(&mem_map, 0);
-        let entry = mem_map.entries[first];
+        let entry = mem_map.entries[first as usize];
         let base = align_address(entry.base, PAGE_ORDER);
         assert!(base <= usize::max_value() as u64);
         assert!(base < entry.base + entry.length);
@@ -171,14 +174,14 @@ impl<'a> FrameAllocator<'a> {
         let addr = self.cur_addr;
         self.cur_addr += PAGE_SIZE;
 
-        let map_entry = self.mem_map.entries[self.cur_map_entry];
+        let map_entry = self.mem_map.entries[self.cur_map_entry as usize];
         let next_addr = self.cur_addr + PAGE_SIZE;
         if (next_addr as u64) > map_entry.base + map_entry.length {
-            self.cur_map_entry = next_entry(&self.mem_map, self.cur_map_entry+1);
+            self.cur_map_entry = next_entry(&self.mem_map, self.cur_map_entry as usize + 1);
             if self.cur_map_entry >= self.mem_map.num_entries {
                 panic!("Out of physical memory.");
             }
-            let next_entry = self.mem_map.entries[self.cur_map_entry];
+            let next_entry = self.mem_map.entries[self.cur_map_entry as usize];
             let new_addr = align_address(next_entry.base, PAGE_ORDER);
             assert!(new_addr <= usize::max_value() as u64);
             assert!(new_addr < next_entry.base + next_entry.length);
